requirements.txt
Django~=2.2.4
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
